初始化文件

+++ src/main/java/com/cloudplatform/utils/FileDeleteUtil.java
@@ -10,19 +10,21 @@ import java.io.File;
  * @Description:
  */
 public class FileDeleteUtil {
-    public static void deleteFile(File file){
+    public static void deleteFile(File file) {
         //取得这个目录下的所有子文件对象
         File[] files = file.listFiles();
         //遍历该目录下的文件对象
-        for (File f: files){
-            //打印文件名
-            String name = file.getName();
-            System.out.println(name);
-            //判断子目录是否存在子目录,如果是文件则删除
-            if (f.isDirectory()){
-                deleteFile(f);
-            }else {
-                f.delete();
+        if (null != files) {
+            for (File f : files) {
+                //打印文件名
+                String name = file.getName();
+                System.out.println(name);
+                //判断子目录是否存在子目录,如果是文件则删除
+                if (f.isDirectory()) {
+                    deleteFile(f);
+                } else {
+                    f.delete();
+                }
             }
         }
         //删除空文件夹  for循环已经把上一层节点的目录清空。
